config_vehicles.hpp cleanup for marines2
Took the marines2 config_vehicles.hpp and split it down by camo.
