+++ src/Addons/34thPRC_AceArsenalExtended/data/aax_models/weapons/armour_standard/odst2/config.hpp
