+++ src/Addons/34thPRC_AceArsenalExtended/data/aax_models/weapons/config.hpp
