+++ src/Addons/34thPRC_ArmourStandard/data/marines2/arctic/config_vehicles.hpp
@@ -1,0 +1,119 @@
+class 34thPRC_ArmourStandard_Marines2_Trooper_Arctic_Light: 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Light
+{
+	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Arctic_Light";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\Arctic\Winter_UpperBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\Arctic\Winter_LowerBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\Arctic\Winter_TrooperCollar.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
+		"34thPRC_ArmourStandard\data\marines2\Arctic\White_UpperArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\Arctic\White_LowerArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\Arctic\White_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\Arctic\White_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\Arctic\White_TrooperShoulders.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Trooper_Arctic_Medium: 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Medium
+{
+	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Arctic_Medium";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\Arctic\Winter_UpperBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\Arctic\Winter_LowerBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\Arctic\Winter_TrooperCollar.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
+		"34thPRC_ArmourStandard\data\marines2\Arctic\White_UpperArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\Arctic\White_LowerArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\Arctic\White_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\Arctic\White_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\Arctic\White_TrooperShoulders.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Trooper_Arctic_Heavy: 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Heavy
+{
+	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Arctic_Heavy";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\Arctic\Winter_UpperBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\Arctic\Winter_LowerBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\Arctic\Winter_TrooperCollar.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
+		"34thPRC_ArmourStandard\data\marines2\Arctic\White_UpperArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\Arctic\White_LowerArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\Arctic\White_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\Arctic\White_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\Arctic\White_TrooperShoulders.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Trooper_Arctic_NS: 34thPRC_ArmourStandard_Marines2_Trooper_Urban_NS
+{
+	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Arctic_NS";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\Arctic\Winter_UpperBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\Arctic\Winter_LowerBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\Arctic\Winter_TrooperCollar.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
+		"34thPRC_ArmourStandard\data\marines2\Arctic\White_UpperArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\Arctic\White_LowerArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\Arctic\White_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\Arctic\White_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\Arctic\White_TrooperShoulders.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Trooper_Arctic_Light_NoCollar: 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Light_NoCollar
+{
+	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Arctic_Light_NoCollar";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\Arctic\Winter_UpperBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\Arctic\Winter_LowerBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\Arctic\Winter_TrooperCollar.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
+		"34thPRC_ArmourStandard\data\marines2\Arctic\White_UpperArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\Arctic\White_LowerArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\Arctic\White_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\Arctic\White_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\Arctic\White_TrooperShoulders.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Trooper_Arctic_Medium_NoCollar: 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Medium_NoCollar
+{
+	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Arctic_Medium_NoCollar";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\Arctic\Winter_UpperBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\Arctic\Winter_LowerBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\Arctic\Winter_TrooperCollar.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
+		"34thPRC_ArmourStandard\data\marines2\Arctic\White_UpperArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\Arctic\White_LowerArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\Arctic\White_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\Arctic\White_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\Arctic\White_TrooperShoulders.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Trooper_Arctic_Heavy_NoCollar: 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Heavy_NoCollar
+{
+	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Arctic_Heavy_NoCollar";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\Arctic\Winter_UpperBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\Arctic\Winter_LowerBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\Arctic\Winter_TrooperCollar.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
+		"34thPRC_ArmourStandard\data\marines2\Arctic\White_UpperArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\Arctic\White_LowerArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\Arctic\White_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\Arctic\White_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\Arctic\White_TrooperShoulders.paa"
+	};
+};
