+++ src/Addons/34thPRC_ArmourStandard/data/marines2/config_vehicles.hpp
@@ -5,732 +5,9 @@ class MA_Marine_Heavy_BDU_NC; // MA_Armor
 class MA_Marine_Light_BDU_NC; // MA_Armor
 class MA_Marine_Medium_BDU_NC; // MA_Armor
 class MA_Marine_NS_BDU; //MA_Armor
-class 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Light: MA_Marine_Light_BDU
-{
-	scope=1;
-	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_UpperBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_LowerBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperCollar.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_UpperArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_LowerArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Medium: MA_Marine_Medium_BDU
-{
-	scope=1;
-	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Urban_Medium";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_UpperBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_LowerBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperCollar.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_UpperArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_LowerArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Heavy: MA_Marine_Heavy_BDU
-{
-	scope=1;
-	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Urban_Heavy";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_UpperBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_LowerBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperCollar.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_UpperArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_LowerArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Trooper_Urban_NS: MA_Marine_NS_BDU
-{
-	scope=1;
-	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Urban_NS";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_UpperBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_LowerBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperCollar.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_UpperArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_LowerArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Light_NoCollar: MA_Marine_Light_BDU_NC
-{
-	scope=1;
-	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light_NoCollar";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_UpperBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_LowerBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperCollar.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_UpperArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_LowerArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Medium_NoCollar: MA_Marine_Medium_BDU_NC
-{
-	scope=1;
-	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Urban_Medium_NoCollar";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_UpperBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_LowerBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperCollar.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_UpperArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_LowerArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Heavy_NoCollar: MA_Marine_Heavy_BDU_NC
-{
-	scope=1;
-	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Urban_Heavy_NoCollar";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_UpperBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_LowerBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperCollar.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_UpperArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_LowerArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Blue_Light: MA_Marine_Light_BDU
-{
-	scope=1;
-	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Urban_Blue_Light";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\urban\Blue_UpperBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Blue_LowerBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Blue_TrooperCollar.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_UpperArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_LowerArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Blue_Medium: MA_Marine_Medium_BDU
-{
-	scope=1;
-	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Urban_Blue_Medium";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\urban\Blue_UpperBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Blue_LowerBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Blue_TrooperCollar.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_UpperArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_LowerArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Blue_Heavy: MA_Marine_Heavy_BDU
-{
-	scope=1;
-	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Urban_Blue_Heavy";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\urban\Blue_UpperBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Blue_LowerBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Blue_TrooperCollar.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_UpperArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_LowerArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Blue_NS: 34thPRC_ArmourStandard_Marines2_Trooper_Urban_NS
-{
-	scope=1;
-	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Urban_Blue_NS";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\urban\Blue_UpperBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Blue_LowerBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Blue_TrooperCollar.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_UpperArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_LowerArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Blue_Light_NoCollar: MA_Marine_Light_BDU_NC
-{
-	scope=1;
-	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Urban_Blue_Light_NoCollar";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\urban\Blue_UpperBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Blue_LowerBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Blue_TrooperCollar.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_UpperArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_LowerArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Blue_Medium_NoCollar: MA_Marine_Medium_BDU_NC
-{
-	scope=1;
-	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Urban_Blue_Medium_NoCollar";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\urban\Blue_UpperBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Blue_LowerBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Blue_TrooperCollar.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_UpperArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_LowerArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Blue_Heavy_NoCollar: MA_Marine_Heavy_BDU_NC
-{
-	scope=1;
-	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Urban_Blue_Heavy_NoCollar";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\urban\Blue_UpperBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Blue_LowerBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Blue_TrooperCollar.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_UpperArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_LowerArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Trooper_Arctic_Light: 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Light
-{
-	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Arctic_Light";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\Arctic\Winter_UpperBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arctic\Winter_LowerBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arctic\Winter_TrooperCollar.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arctic\White_UpperArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arctic\White_LowerArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arctic\White_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arctic\White_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arctic\White_TrooperShoulders.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Trooper_Arctic_Medium: 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Medium
-{
-	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Arctic_Medium";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\Arctic\Winter_UpperBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arctic\Winter_LowerBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arctic\Winter_TrooperCollar.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arctic\White_UpperArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arctic\White_LowerArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arctic\White_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arctic\White_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arctic\White_TrooperShoulders.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Trooper_Arctic_Heavy: 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Heavy
-{
-	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Arctic_Heavy";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\Arctic\Winter_UpperBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arctic\Winter_LowerBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arctic\Winter_TrooperCollar.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arctic\White_UpperArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arctic\White_LowerArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arctic\White_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arctic\White_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arctic\White_TrooperShoulders.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Trooper_Arctic_NS: 34thPRC_ArmourStandard_Marines2_Trooper_Urban_NS
-{
-	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Arctic_NS";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\Arctic\Winter_UpperBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arctic\Winter_LowerBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arctic\Winter_TrooperCollar.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arctic\White_UpperArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arctic\White_LowerArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arctic\White_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arctic\White_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arctic\White_TrooperShoulders.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Trooper_Arctic_Light_NoCollar: 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Light_NoCollar
-{
-	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Arctic_Light_NoCollar";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\Arctic\Winter_UpperBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arctic\Winter_LowerBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arctic\Winter_TrooperCollar.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arctic\White_UpperArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arctic\White_LowerArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arctic\White_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arctic\White_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arctic\White_TrooperShoulders.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Trooper_Arctic_Medium_NoCollar: 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Medium_NoCollar
-{
-	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Arctic_Medium_NoCollar";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\Arctic\Winter_UpperBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arctic\Winter_LowerBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arctic\Winter_TrooperCollar.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arctic\White_UpperArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arctic\White_LowerArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arctic\White_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arctic\White_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arctic\White_TrooperShoulders.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Trooper_Arctic_Heavy_NoCollar: 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Heavy_NoCollar
-{
-	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Arctic_Heavy_NoCollar";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\Arctic\Winter_UpperBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arctic\Winter_LowerBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arctic\Winter_TrooperCollar.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arctic\White_UpperArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arctic\White_LowerArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arctic\White_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arctic\White_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arctic\White_TrooperShoulders.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Trooper_Arid_Light: 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Light
-{
-	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Arid_Light";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\Arid\Desert_UpperBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arid\Desert_LowerBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arid\Desert_TrooperCollar.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arid\Tan_UpperArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arid\Tan_LowerArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arid\Tan_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arid\Tan_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arid\Tan_TrooperShoulders.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Trooper_Arid_Medium: 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Medium
-{
-	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Arid_Medium";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\Arid\Desert_UpperBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arid\Desert_LowerBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arid\Desert_TrooperCollar.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arid\Tan_UpperArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arid\Tan_LowerArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arid\Tan_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arid\Tan_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arid\Tan_TrooperShoulders.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Trooper_Arid_Heavy: 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Heavy
-{
-	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Arid_Heavy";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\Arid\Desert_UpperBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arid\Desert_LowerBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arid\Desert_TrooperCollar.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arid\Tan_UpperArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arid\Tan_LowerArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arid\Tan_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arid\Tan_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arid\Tan_TrooperShoulders.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Trooper_Arid_NS: 34thPRC_ArmourStandard_Marines2_Trooper_Urban_NS
-{
-	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Arid_NS";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\Arid\Desert_UpperBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arid\Desert_LowerBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arid\Desert_TrooperCollar.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arid\Tan_UpperArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arid\Tan_LowerArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arid\Tan_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arid\Tan_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arid\Tan_TrooperShoulders.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Trooper_Arid_Light_NoCollar: 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Light_NoCollar
-{
-	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Arid_Light_NoCollar";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\Arid\Desert_UpperBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arid\Desert_LowerBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arid\Desert_TrooperCollar.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arid\Tan_UpperArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arid\Tan_LowerArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arid\Tan_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arid\Tan_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arid\Tan_TrooperShoulders.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Trooper_Arid_Medium_NoCollar: 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Medium_NoCollar
-{
-	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Arid_Medium_NoCollar";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\Arid\Desert_UpperBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arid\Desert_LowerBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arid\Desert_TrooperCollar.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arid\Tan_UpperArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arid\Tan_LowerArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arid\Tan_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arid\Tan_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arid\Tan_TrooperShoulders.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Trooper_Arid_Heavy_NoCollar: 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Heavy_NoCollar
-{
-	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Arid_Heavy_NoCollar";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\Arid\Desert_UpperBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arid\Desert_LowerBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arid\Desert_TrooperCollar.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arid\Tan_UpperArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arid\Tan_LowerArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arid\Tan_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arid\Tan_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Arid\Tan_TrooperShoulders.paa"
-	};
-};
 
-class 34thPRC_ArmourStandard_Marines2_Trooper_Tropic_Light: 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Light
-{
-	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Tropic_Light";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_UpperBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_LowerBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperCollar.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_UpperArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_LowerArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperShoulders.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Trooper_Tropic_Medium: 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Medium
-{
-	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Tropic_Medium";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_UpperBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_LowerBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperCollar.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_UpperArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_LowerArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperShoulders.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Trooper_Tropic_Heavy: 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Heavy
-{
-	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Tropic_Heavy";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_UpperBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_LowerBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperCollar.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_UpperArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_LowerArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperShoulders.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Trooper_Tropic_NS: 34thPRC_ArmourStandard_Marines2_Trooper_Urban_NS
-{
-	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Tropic_NS";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_UpperBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_LowerBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperCollar.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_UpperArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_LowerArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperShoulders.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Trooper_Tropic_Light_NoCollar: 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Light_NoCollar
-{
-	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Tropic_Light_NoCollar";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_UpperBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_LowerBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperCollar.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_UpperArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_LowerArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperShoulders.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Trooper_Tropic_Medium_NoCollar: 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Medium_NoCollar
-{
-	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Tropic_Medium_NoCollar";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_UpperBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_LowerBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperCollar.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_UpperArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_LowerArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperShoulders.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Trooper_Tropic_Heavy_NoCollar: 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Heavy_NoCollar
-{
-	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Tropic_Heavy_NoCollar";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_UpperBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_LowerBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperCollar.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_UpperArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_LowerArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperShoulders.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Trooper_Woodland_Light: 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Light
-{
-	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Woodland_Light";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Woodland_UpperBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Woodland_LowerBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Woodland_TrooperCollar.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_UpperArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_LowerArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_TrooperShoulders.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Trooper_Woodland_Medium: 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Medium
-{
-	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Woodland_Medium";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Woodland_UpperBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Woodland_LowerBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Woodland_TrooperCollar.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_UpperArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_LowerArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_TrooperShoulders.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Trooper_Woodland_Heavy: 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Heavy
-{
-	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Woodland_Heavy";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Woodland_UpperBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Woodland_LowerBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Woodland_TrooperCollar.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_UpperArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_LowerArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_TrooperShoulders.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Trooper_Woodland_NS: 34thPRC_ArmourStandard_Marines2_Trooper_Urban_NS
-{
-	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Woodland_NS";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Woodland_UpperBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Woodland_LowerBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Woodland_TrooperCollar.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_UpperArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_LowerArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_TrooperShoulders.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Trooper_Woodland_Light_NoCollar: 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Light_NoCollar
-{
-	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Woodland_Light_NoCollar";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Woodland_UpperBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Woodland_LowerBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Woodland_TrooperCollar.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_UpperArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_LowerArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_TrooperShoulders.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Trooper_Woodland_Medium_NoCollar: 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Medium_NoCollar
-{
-	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Woodland_Medium_NoCollar";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Woodland_UpperBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Woodland_LowerBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Woodland_TrooperCollar.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_UpperArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_LowerArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_TrooperShoulders.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Trooper_Woodland_Heavy_NoCollar: 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Heavy_NoCollar
-{
-	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Woodland_Heavy_NoCollar";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Woodland_UpperBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Woodland_LowerBDU.paa",
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Woodland_TrooperCollar.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
-		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_UpperArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_LowerArmor.paa",
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_TrooperShoulders.paa",
-		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_TrooperShoulders.paa"
-	};
-};
+#include "urban\config_vehicles.hpp"
+#include "arid\config_vehicles.hpp"
+#include "arctic\config_vehicles.hpp"
+#include "tropic\config_vehicles.hpp"
+#include "woodland\config_vehicles.hpp"
