+++ src/Addons/34thPRC_ArmourStandard/data/marines2/tropic/config_vehicles.hpp
@@ -1,0 +1,119 @@
+class 34thPRC_ArmourStandard_Marines2_Trooper_Tropic_Light: 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Light
+{
+	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Tropic_Light";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_UpperBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_LowerBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperCollar.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_UpperArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_LowerArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperShoulders.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Trooper_Tropic_Medium: 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Medium
+{
+	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Tropic_Medium";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_UpperBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_LowerBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperCollar.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_UpperArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_LowerArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperShoulders.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Trooper_Tropic_Heavy: 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Heavy
+{
+	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Tropic_Heavy";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_UpperBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_LowerBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperCollar.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_UpperArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_LowerArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperShoulders.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Trooper_Tropic_NS: 34thPRC_ArmourStandard_Marines2_Trooper_Urban_NS
+{
+	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Tropic_NS";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_UpperBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_LowerBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperCollar.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_UpperArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_LowerArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperShoulders.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Trooper_Tropic_Light_NoCollar: 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Light_NoCollar
+{
+	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Tropic_Light_NoCollar";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_UpperBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_LowerBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperCollar.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_UpperArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_LowerArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperShoulders.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Trooper_Tropic_Medium_NoCollar: 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Medium_NoCollar
+{
+	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Tropic_Medium_NoCollar";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_UpperBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_LowerBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperCollar.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_UpperArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_LowerArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperShoulders.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Trooper_Tropic_Heavy_NoCollar: 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Heavy_NoCollar
+{
+	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Tropic_Heavy_NoCollar";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_UpperBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_LowerBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperCollar.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_UpperArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_LowerArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\Tropic\Brown_TrooperShoulders.paa"
+	};
+};
