+++ src/Addons/34thPRC_ArmourStandard/data/marines2/urban/config_vehicles.hpp
@@ -1,0 +1,252 @@
+class 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Light: MA_Marine_Light_BDU
+{
+	scope=1;
+	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_UpperBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_LowerBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperCollar.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_UpperArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_LowerArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Medium: MA_Marine_Medium_BDU
+{
+	scope=1;
+	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Urban_Medium";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_UpperBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_LowerBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperCollar.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_UpperArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_LowerArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Heavy: MA_Marine_Heavy_BDU
+{
+	scope=1;
+	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Urban_Heavy";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_UpperBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_LowerBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperCollar.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_UpperArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_LowerArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Trooper_Urban_NS: MA_Marine_NS_BDU
+{
+	scope=1;
+	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Urban_NS";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_UpperBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_LowerBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperCollar.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_UpperArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_LowerArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Light_NoCollar: MA_Marine_Light_BDU_NC
+{
+	scope=1;
+	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light_NoCollar";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_UpperBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_LowerBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperCollar.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_UpperArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_LowerArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Medium_NoCollar: MA_Marine_Medium_BDU_NC
+{
+	scope=1;
+	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Urban_Medium_NoCollar";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_UpperBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_LowerBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperCollar.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_UpperArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_LowerArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Heavy_NoCollar: MA_Marine_Heavy_BDU_NC
+{
+	scope=1;
+	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Urban_Heavy_NoCollar";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_UpperBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_LowerBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperCollar.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_UpperArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_LowerArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Blue_Light: MA_Marine_Light_BDU
+{
+	scope=1;
+	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Urban_Blue_Light";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\urban\Blue_UpperBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Blue_LowerBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Blue_TrooperCollar.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_UpperArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_LowerArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Blue_Medium: MA_Marine_Medium_BDU
+{
+	scope=1;
+	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Urban_Blue_Medium";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\urban\Blue_UpperBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Blue_LowerBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Blue_TrooperCollar.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_UpperArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_LowerArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Blue_Heavy: MA_Marine_Heavy_BDU
+{
+	scope=1;
+	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Urban_Blue_Heavy";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\urban\Blue_UpperBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Blue_LowerBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Blue_TrooperCollar.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_UpperArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_LowerArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Blue_NS: 34thPRC_ArmourStandard_Marines2_Trooper_Urban_NS
+{
+	scope=1;
+	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Urban_Blue_NS";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\urban\Blue_UpperBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Blue_LowerBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Blue_TrooperCollar.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_UpperArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_LowerArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Blue_Light_NoCollar: MA_Marine_Light_BDU_NC
+{
+	scope=1;
+	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Urban_Blue_Light_NoCollar";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\urban\Blue_UpperBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Blue_LowerBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Blue_TrooperCollar.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_UpperArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_LowerArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Blue_Medium_NoCollar: MA_Marine_Medium_BDU_NC
+{
+	scope=1;
+	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Urban_Blue_Medium_NoCollar";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\urban\Blue_UpperBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Blue_LowerBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Blue_TrooperCollar.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_UpperArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_LowerArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Blue_Heavy_NoCollar: MA_Marine_Heavy_BDU_NC
+{
+	scope=1;
+	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Urban_Blue_Heavy_NoCollar";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\urban\Blue_UpperBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Blue_LowerBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Blue_TrooperCollar.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_UpperArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_LowerArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa"
+	};
+};
