+++ src/Addons/34thPRC_ArmourStandard/data/marines2/woodland/config_vehicles.hpp
@@ -1,0 +1,119 @@
+class 34thPRC_ArmourStandard_Marines2_Trooper_Woodland_Light: 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Light
+{
+	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Woodland_Light";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Woodland_UpperBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Woodland_LowerBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Woodland_TrooperCollar.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_UpperArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_LowerArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_TrooperShoulders.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Trooper_Woodland_Medium: 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Medium
+{
+	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Woodland_Medium";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Woodland_UpperBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Woodland_LowerBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Woodland_TrooperCollar.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_UpperArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_LowerArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_TrooperShoulders.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Trooper_Woodland_Heavy: 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Heavy
+{
+	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Woodland_Heavy";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Woodland_UpperBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Woodland_LowerBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Woodland_TrooperCollar.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_UpperArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_LowerArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_TrooperShoulders.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Trooper_Woodland_NS: 34thPRC_ArmourStandard_Marines2_Trooper_Urban_NS
+{
+	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Woodland_NS";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Woodland_UpperBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Woodland_LowerBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Woodland_TrooperCollar.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_UpperArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_LowerArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_TrooperShoulders.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Trooper_Woodland_Light_NoCollar: 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Light_NoCollar
+{
+	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Woodland_Light_NoCollar";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Woodland_UpperBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Woodland_LowerBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Woodland_TrooperCollar.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_UpperArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_LowerArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_TrooperShoulders.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Trooper_Woodland_Medium_NoCollar: 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Medium_NoCollar
+{
+	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Woodland_Medium_NoCollar";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Woodland_UpperBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Woodland_LowerBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Woodland_TrooperCollar.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_UpperArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_LowerArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_TrooperShoulders.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Trooper_Woodland_Heavy_NoCollar: 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Heavy_NoCollar
+{
+	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Woodland_Heavy_NoCollar";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Woodland_UpperBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Woodland_LowerBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Woodland_TrooperCollar.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_UpperArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_LowerArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\Woodland\Green_TrooperShoulders.paa"
+	};
+};
